+++ libs/react/src/lib/accordion/accordion-item.tsx
@@ -20,13 +20,14 @@ export interface AccordionItemInterface {
   onClick?: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void
   onOpen?: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void
   onClose?: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void
+  defaultOpen?: boolean
 }
 
 const AccordionItem = ({ item, index, uuid }: AccordionItemProps) => {
   const { labelElementLevel, label, subLabel, content, customLabel } = item
   const contentRef: RefObject<HTMLDivElement> = useRef<HTMLDivElement>(null)
 
-  const [isOpen, setIsOpen] = useState(false)
+  const [isOpen, setIsOpen] = useState(item.defaultOpen || false)
 
   const handleOnClick = (
     event: React.MouseEvent<HTMLButtonElement, MouseEvent>,