+++ libs/core/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## 1.31.0
+
+### Minor Changes
+
+- 8b92d27: **Calendar:** Add ability to customize aria-label for dates
+- 8b92d27: **Calendar:** Add ability to cancel focus event
+- 8b92d27: **Calender:** Add option to hide day name header
+- 8b92d27: **Calendar:** Add option to hide days outside of current month
+
+## 1.30.0
+
+### Minor Changes
+
+- f302973: **Popover:** Add options for modality, backdrop and positioning
+
+### Patch Changes
+
+- f302973: **Container:** Default to `display: block`
+
+## 1.29.0
+
+### Minor Changes
+
+- 663c34d: **Segmented Control:** Add `disabled` property to `gds-segment`
+
 ## 1.28.1
 
 ### Patch Changes