+++ libs/react/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 3.10.0
+
+### Minor Changes
+
+- 525b2c6: **Accordion:** Add `defaultOpen` prop
+
+## 3.9.0
+
+### Minor Changes
+
+- 663c34d: **Segmented Control:** Add new React wrapper
+
+### Patch Changes
+
+- Updated dependencies [663c34d]
+  - @sebgroup/green-core@1.29.0
+
 ## 3.8.1
 
 ### Patch Changes